refactor(types): keep only type exports, move constants to models and client
- Remove all constant exports from @soker90/finper-types (now type-only)
- Define constants canonically in @soker90/finper-models model files
- Add BUDGET_RULE_CLASS to categories model and models index
- Create local enums.ts in client constants for client-side usage
- Update all 22 API source files to import from @soker90/finper-models
- Update client types barrel to import constants from local constants/enums
- types package now safe as devDependency (zero runtime values)
