+++ packages/api/src/services/budget.service.ts
@@ -1,4 +1,4 @@
-import { TRANSACTION, TransactionType } from '@soker90/finper-types'
+import { TRANSACTION, TransactionType } from '@soker90/finper-models'
 import { CategoryModel, mongoose, IBudget, BudgetModel } from '@soker90/finper-models'
 import { calcBudgetByMonths, getTransactionsSumByMonth } from './utils'
