+++ packages/api/src/services/stats.service.ts
@@ -1,4 +1,4 @@
-import { TRANSACTION } from '@soker90/finper-types'
+import { TRANSACTION } from '@soker90/finper-models'
 import { TransactionModel } from '@soker90/finper-models'
 import { TagSummary, TagHistoric, TagDetail, TagYearSummary } from '../types/stats.types'
 import { roundNumber } from '../utils/roundNumber'
