+++ packages/api/src/services/utils/getTransactionsAmount.ts
@@ -1,4 +1,4 @@
-import { TRANSACTION } from '@soker90/finper-types'
+import { TRANSACTION } from '@soker90/finper-models'
 import { ITransaction } from '@soker90/finper-models'
 
 export const getTransactionAmount = (transaction: ITransaction): number =>
