+++ packages/api/src/validators/goal/validate-goal-create-params.ts
@@ -1,6 +1,6 @@
 import Joi from 'joi'
 import Boom from '@hapi/boom'
-import { GOAL_COLORS, GOAL_ICONS } from '@soker90/finper-types'
+import { GOAL_COLORS, GOAL_ICONS } from '@soker90/finper-models'
 
 export const validateGoalCreateParams = async (data: Record<string, unknown>) => {
   const schema = Joi.object({
