+++ packages/api/src/validators/stock/validate-stock-create-params.ts
@@ -1,6 +1,6 @@
 import Joi from 'joi'
 import Boom from '@hapi/boom'
-import { STOCK_TYPE } from '@soker90/finper-types'
+import { STOCK_TYPE } from '@soker90/finper-models'
 import { IStock } from '@soker90/finper-models'
 
 export const validateStockCreateParams = async (body: Omit<IStock, 'user'>) => {
