+++ packages/api/src/validators/supply/validate-supply-create-params.ts
@@ -1,6 +1,6 @@
 import Joi from 'joi'
 import Boom from '@hapi/boom'
-import { SUPPLY_TYPE } from '@soker90/finper-types'
+import { SUPPLY_TYPE } from '@soker90/finper-models'
 import { validatePropertyExist } from '../property'
 
 export const validateSupplyCreateParams = async (data: Record<string, string>) => {
