+++ packages/api/src/validators/transaction/validate-transaction-edit-params.ts
@@ -1,6 +1,6 @@
 import Joi from 'joi'
 import Boom from '@hapi/boom'
-import { TRANSACTION } from '@soker90/finper-types'
+import { TRANSACTION } from '@soker90/finper-models'
 import { ITransaction } from '@soker90/finper-models'
 import { validateTransactionExist } from './validate-transaction-exist'
