+++ packages/models/src/models/goals.ts
@@ -1,7 +1,32 @@
 import { Schema, model, HydratedDocument } from 'mongoose'
-import { GOAL_COLORS, GOAL_ICONS, type GoalColor, type GoalIcon } from '@soker90/finper-types'
+import type { GoalColor, GoalIcon } from '@soker90/finper-types'
 
-export { GOAL_COLORS, GOAL_ICONS }
+export const GOAL_COLORS = [
+  '#4CAF50',
+  '#2196F3',
+  '#9C27B0',
+  '#FF9800',
+  '#F44336',
+  '#00BCD4',
+  '#795548',
+  '#607D8B',
+  '#E91E63',
+  '#FFC107',
+] as const
+
+export const GOAL_ICONS = [
+  'DollarOutlined',
+  'HomeOutlined',
+  'CarOutlined',
+  'LaptopOutlined',
+  'HeartOutlined',
+  'RocketOutlined',
+  'GiftOutlined',
+  'BankOutlined',
+  'TrophyOutlined',
+  'StarOutlined',
+] as const
+
 export type { GoalColor, GoalIcon }
 
 export interface IGoal {
