+++ packages/models/src/models/loan-payments.ts
@@ -1,7 +1,11 @@
 import { Schema, model, HydratedDocument, Types } from 'mongoose'
-import { LOAN_PAYMENT, type LoanPaymentType } from '@soker90/finper-types'
+import type { LoanPaymentType } from '@soker90/finper-types'
 
-export { LOAN_PAYMENT }
+export const LOAN_PAYMENT = {
+  ORDINARY: 'ordinary',
+  EXTRAORDINARY: 'extraordinary',
+} as const
+
 export type { LoanPaymentType }
 
 export interface ILoanPayment {
