+++ packages/models/src/models/transactions.ts
@@ -1,7 +1,12 @@
 import { Schema, model, Types, HydratedDocument } from 'mongoose'
-import { TRANSACTION, type TransactionType } from '@soker90/finper-types'
+import type { TransactionType } from '@soker90/finper-types'
 
-export { TRANSACTION }
+export const TRANSACTION = {
+  Expense: 'expense',
+  Income: 'income',
+  NotComputable: 'not_computable',
+} as const
+
 export type { TransactionType }
 
 export type TransactionDocument = HydratedDocument<ITransaction>
