+++ packages/types/src/enums/debt.ts
@@ -1,6 +1,1 @@
-export const DEBT = {
-  FROM: 'from',
-  TO: 'to',
-} as const
-
-export type DebtType = typeof DEBT[keyof typeof DEBT]
+export type DebtType = 'from' | 'to'
