+++ packages/types/src/enums/goal.ts
@@ -1,29 +1,2 @@
-export const GOAL_COLORS = [
-  '#4CAF50',
-  '#2196F3',
-  '#9C27B0',
-  '#FF9800',
-  '#F44336',
-  '#00BCD4',
-  '#795548',
-  '#607D8B',
-  '#E91E63',
-  '#FFC107',
-] as const
-
-export type GoalColor = typeof GOAL_COLORS[number]
-
-export const GOAL_ICONS = [
-  'DollarOutlined',
-  'HomeOutlined',
-  'CarOutlined',
-  'LaptopOutlined',
-  'HeartOutlined',
-  'RocketOutlined',
-  'GiftOutlined',
-  'BankOutlined',
-  'TrophyOutlined',
-  'StarOutlined',
-] as const
-
-export type GoalIcon = typeof GOAL_ICONS[number]
+export type GoalColor = '#4CAF50' | '#2196F3' | '#9C27B0' | '#FF9800' | '#F44336' | '#00BCD4' | '#795548' | '#607D8B' | '#E91E63' | '#FFC107'
+export type GoalIcon = 'DollarOutlined' | 'HomeOutlined' | 'CarOutlined' | 'LaptopOutlined' | 'HeartOutlined' | 'RocketOutlined' | 'GiftOutlined' | 'BankOutlined' | 'TrophyOutlined' | 'StarOutlined'
