+++ packages/types/src/enums/loan.ts
@@ -1,6 +1,1 @@
-export const LOAN_PAYMENT = {
-  ORDINARY: 'ordinary',
-  EXTRAORDINARY: 'extraordinary',
-} as const
-
-export type LoanPaymentType = typeof LOAN_PAYMENT[keyof typeof LOAN_PAYMENT]
+export type LoanPaymentType = 'ordinary' | 'extraordinary'
