+++ packages/types/src/enums/stock.ts
@@ -1,7 +1,1 @@
-export const STOCK_TYPE = {
-  Buy: 'buy',
-  Sell: 'sell',
-  Dividend: 'dividend',
-} as const
-
-export type StockType = typeof STOCK_TYPE[keyof typeof STOCK_TYPE]
+export type StockType = 'buy' | 'sell' | 'dividend'
