+++ packages/types/src/enums/transaction.ts
@@ -1,7 +1,1 @@
-export const TRANSACTION = {
-  Expense: 'expense',
-  Income: 'income',
-  NotComputable: 'not_computable',
-} as const
-
-export type TransactionType = typeof TRANSACTION[keyof typeof TRANSACTION]
+export type TransactionType = 'expense' | 'income' | 'not_computable'
