+++ packages/types/src/index.ts
@@ -1,20 +1,7 @@
-export { TRANSACTION } from './enums/transaction'
 export type { TransactionType } from './enums/transaction'
-
-export { DEBT } from './enums/debt'
 export type { DebtType } from './enums/debt'
-
-export { LOAN_PAYMENT } from './enums/loan'
 export type { LoanPaymentType } from './enums/loan'
-
-export { SUPPLY_TYPE } from './enums/supply'
 export type { SupplyType } from './enums/supply'
-
-export { STOCK_TYPE } from './enums/stock'
 export type { StockType } from './enums/stock'
-
-export { GOAL_COLORS, GOAL_ICONS } from './enums/goal'
 export type { GoalColor, GoalIcon } from './enums/goal'
-
-export { BUDGET_RULE_CLASS } from './enums/budget-rule'
 export type { BudgetRuleClassType } from './enums/budget-rule'
